// src/app/main/models/amplifier.model.ts
export interface Amplifier {
  action: string;
  country: string;
  fullLegalName: string;
  shortName: string;
  status: string;
  industry: string;
  pricingSource: string;
  location: string;
  orgId: string;
  countryOfLocation: string;
  privateKey: string;
  publicKey: string;
  ticker: string;
}